multi object model import (.obj)

+++ src/Rendering/Buffers/IndexBuffer.cpp
@@ -1,9 +1,9 @@
 ﻿#include "IndexBuffer.h"
 
-IndexBuffer::IndexBuffer(GLubyte* pIndices, const unsigned int numIndices)
+IndexBuffer::IndexBuffer(unsigned int* pIndices, const unsigned int numIndices)
 {
     glGenBuffers(1, &_indexBufferID);
-    _pIndices = pIndices;
+    _pIndices   = pIndices;
     _numIndices = numIndices;
 }
 
@@ -12,15 +12,12 @@ IndexBuffer::~IndexBuffer()
     glDeleteBuffers(GL_ELEMENT_ARRAY_BUFFER, &_indexBufferID);
 }
 
-unsigned int IndexBuffer::GetNumIndices() const
-{
-    return _numIndices;
-}
+unsigned int IndexBuffer::GetNumIndices() const { return _numIndices; }
 
 void IndexBuffer::Bind() const
 {
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, _indexBufferID);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, _numIndices * sizeof(GLubyte), _pIndices, GL_STATIC_DRAW);
+    glBufferData(GL_ELEMENT_ARRAY_BUFFER, _numIndices * sizeof(unsigned int), _pIndices, GL_STATIC_DRAW);
 }
 
 void IndexBuffer::Unbind()
