+++ src/Rendering/Buffers/IndexBuffer.h
@@ -5,12 +5,12 @@ class IndexBuffer
 {
     private:
         GLuint       _indexBufferID = 0;
-        GLubyte*     _pIndices;
+        unsigned int*      _pIndices;
         unsigned int _numIndices;
     public:
-        IndexBuffer(GLubyte* pIndices, unsigned int numIndices);
+        IndexBuffer(unsigned int* pIndices, unsigned int numIndices);
         ~IndexBuffer();
         unsigned int GetNumIndices() const;
-        void Bind() const;
-        static void Unbind();
+        void         Bind() const;
+        static void  Unbind();
 };
